Move Validate into {ColumnFamily,DB}Options from ColumnFamiyData/DBImpl
