+++ db/column_family.cc
@@ -116,85 +116,6 @@ void GetIntTblPropCollectorFactory(
   }
 }
 
-Status CheckCompressionSupported(const ColumnFamilyOptions& cf_options) {
-  if (!cf_options.compression_per_level.empty()) {
-    for (size_t level = 0; level < cf_options.compression_per_level.size();
-         ++level) {
-      if (!CompressionTypeSupported(cf_options.compression_per_level[level])) {
-        return Status::InvalidArgument(
-            "Compression type " +
-            CompressionTypeToString(cf_options.compression_per_level[level]) +
-            " is not linked with the binary.");
-      }
-    }
-  } else {
-    if (!CompressionTypeSupported(cf_options.compression)) {
-      return Status::InvalidArgument(
-          "Compression type " +
-          CompressionTypeToString(cf_options.compression) +
-          " is not linked with the binary.");
-    }
-  }
-  if (cf_options.compression_opts.zstd_max_train_bytes > 0) {
-    if (!ZSTD_TrainDictionarySupported()) {
-      return Status::InvalidArgument(
-          "zstd dictionary trainer cannot be used because ZSTD 1.1.3+ "
-          "is not linked with the binary.");
-    }
-    if (cf_options.compression_opts.max_dict_bytes == 0) {
-      return Status::InvalidArgument(
-          "The dictionary size limit (`CompressionOptions::max_dict_bytes`) "
-          "should be nonzero if we're using zstd's dictionary generator.");
-    }
-  }
-
-  if (!CompressionTypeSupported(cf_options.blob_compression_type)) {
-    std::ostringstream oss;
-    oss << "The specified blob compression type "
-        << CompressionTypeToString(cf_options.blob_compression_type)
-        << " is not available.";
-
-    return Status::InvalidArgument(oss.str());
-  }
-
-  return Status::OK();
-}
-
-Status CheckConcurrentWritesSupported(const ColumnFamilyOptions& cf_options) {
-  if (cf_options.inplace_update_support) {
-    return Status::InvalidArgument(
-        "In-place memtable updates (inplace_update_support) is not compatible "
-        "with concurrent writes (allow_concurrent_memtable_write)");
-  }
-  if (!cf_options.memtable_factory->IsInsertConcurrentlySupported()) {
-    return Status::InvalidArgument(
-        "Memtable doesn't concurrent writes (allow_concurrent_memtable_write)");
-  }
-  return Status::OK();
-}
-
-Status CheckCFPathsSupported(const DBOptions& db_options,
-                             const ColumnFamilyOptions& cf_options) {
-  // More than one cf_paths are supported only in universal
-  // and level compaction styles. This function also checks the case
-  // in which cf_paths is not specified, which results in db_paths
-  // being used.
-  if ((cf_options.compaction_style != kCompactionStyleUniversal) &&
-      (cf_options.compaction_style != kCompactionStyleLevel)) {
-    if (cf_options.cf_paths.size() > 1) {
-      return Status::NotSupported(
-          "More than one CF paths are only supported in "
-          "universal and level compaction styles. ");
-    } else if (cf_options.cf_paths.empty() &&
-               db_options.db_paths.size() > 1) {
-      return Status::NotSupported(
-          "More than one DB paths are only supported in "
-          "universal and level compaction styles. ");
-    }
-  }
-  return Status::OK();
-}
-
 namespace {
 const uint64_t kDefaultTtl = 0xfffffffffffffffe;
 const uint64_t kDefaultPeriodicCompSecs = 0xfffffffffffffffe;
@@ -1315,67 +1236,6 @@ void ColumnFamilyData::ResetThreadLocalSuperVersions() {
     // unref'ing super_version_.
     assert(!was_last_ref);
   }
-}
-
-Status ColumnFamilyData::ValidateOptions(
-    const DBOptions& db_options, const ColumnFamilyOptions& cf_options) {
-  Status s;
-  s = CheckCompressionSupported(cf_options);
-  if (s.ok() && db_options.allow_concurrent_memtable_write) {
-    s = CheckConcurrentWritesSupported(cf_options);
-  }
-  if (s.ok() && db_options.unordered_write &&
-      cf_options.max_successive_merges != 0) {
-    s = Status::InvalidArgument(
-        "max_successive_merges > 0 is incompatible with unordered_write");
-  }
-  if (s.ok()) {
-    s = CheckCFPathsSupported(db_options, cf_options);
-  }
-  if (!s.ok()) {
-    return s;
-  }
-
-  if (cf_options.ttl > 0 && cf_options.ttl != kDefaultTtl) {
-    if (!cf_options.table_factory->IsInstanceOf(
-            TableFactory::kBlockBasedTableName())) {
-      return Status::NotSupported(
-          "TTL is only supported in Block-Based Table format. ");
-    }
-  }
-
-  if (cf_options.periodic_compaction_seconds > 0 &&
-      cf_options.periodic_compaction_seconds != kDefaultPeriodicCompSecs) {
-    if (!cf_options.table_factory->IsInstanceOf(
-            TableFactory::kBlockBasedTableName())) {
-      return Status::NotSupported(
-          "Periodic Compaction is only supported in "
-          "Block-Based Table format. ");
-    }
-  }
-
-  if (cf_options.enable_blob_garbage_collection) {
-    if (cf_options.blob_garbage_collection_age_cutoff < 0.0 ||
-        cf_options.blob_garbage_collection_age_cutoff > 1.0) {
-      return Status::InvalidArgument(
-          "The age cutoff for blob garbage collection should be in the range "
-          "[0.0, 1.0].");
-    }
-    if (cf_options.blob_garbage_collection_force_threshold < 0.0 ||
-        cf_options.blob_garbage_collection_force_threshold > 1.0) {
-      return Status::InvalidArgument(
-          "The garbage ratio threshold for forcing blob garbage collection "
-          "should be in the range [0.0, 1.0].");
-    }
-  }
-
-  if (cf_options.compaction_style == kCompactionStyleFIFO &&
-      db_options.max_open_files != -1 && cf_options.ttl > 0) {
-    return Status::NotSupported(
-        "FIFO compaction only supported with max_open_files = -1.");
-  }
-
-  return s;
 }
 
 #ifndef ROCKSDB_LITE
