+++ db/column_family_test.cc
@@ -14,6 +14,7 @@
 
 #include "db/db_impl/db_impl.h"
 #include "db/db_test_util.h"
+#include "options/options_helper.h"
 #include "options/options_parser.h"
 #include "port/port.h"
 #include "port/stack_trace.h"
@@ -3390,21 +3391,19 @@ TEST(ColumnFamilyTest, ValidateBlobGCCutoff) {
   cf_options.enable_blob_garbage_collection = true;
 
   cf_options.blob_garbage_collection_age_cutoff = -0.5;
-  ASSERT_TRUE(ColumnFamilyData::ValidateOptions(db_options, cf_options)
-                  .IsInvalidArgument());
+  ASSERT_TRUE(ValidateOptions(db_options, cf_options).IsInvalidArgument());
 
   cf_options.blob_garbage_collection_age_cutoff = 0.0;
-  ASSERT_OK(ColumnFamilyData::ValidateOptions(db_options, cf_options));
+  ASSERT_OK(ValidateOptions(db_options, cf_options));
 
   cf_options.blob_garbage_collection_age_cutoff = 0.5;
-  ASSERT_OK(ColumnFamilyData::ValidateOptions(db_options, cf_options));
+  ASSERT_OK(ValidateOptions(db_options, cf_options));
 
   cf_options.blob_garbage_collection_age_cutoff = 1.0;
-  ASSERT_OK(ColumnFamilyData::ValidateOptions(db_options, cf_options));
+  ASSERT_OK(ValidateOptions(db_options, cf_options));
 
   cf_options.blob_garbage_collection_age_cutoff = 1.5;
-  ASSERT_TRUE(ColumnFamilyData::ValidateOptions(db_options, cf_options)
-                  .IsInvalidArgument());
+  ASSERT_TRUE(ValidateOptions(db_options, cf_options).IsInvalidArgument());
 }
 
 TEST(ColumnFamilyTest, ValidateBlobGCForceThreshold) {
@@ -3414,21 +3413,19 @@ TEST(ColumnFamilyTest, ValidateBlobGCForceThreshold) {
   cf_options.enable_blob_garbage_collection = true;
 
   cf_options.blob_garbage_collection_force_threshold = -0.5;
-  ASSERT_TRUE(ColumnFamilyData::ValidateOptions(db_options, cf_options)
-                  .IsInvalidArgument());
+  ASSERT_TRUE(ValidateOptions(db_options, cf_options).IsInvalidArgument());
 
   cf_options.blob_garbage_collection_force_threshold = 0.0;
-  ASSERT_OK(ColumnFamilyData::ValidateOptions(db_options, cf_options));
+  ASSERT_OK(ValidateOptions(db_options, cf_options));
 
   cf_options.blob_garbage_collection_force_threshold = 0.5;
-  ASSERT_OK(ColumnFamilyData::ValidateOptions(db_options, cf_options));
+  ASSERT_OK(ValidateOptions(db_options, cf_options));
 
   cf_options.blob_garbage_collection_force_threshold = 1.0;
-  ASSERT_OK(ColumnFamilyData::ValidateOptions(db_options, cf_options));
+  ASSERT_OK(ValidateOptions(db_options, cf_options));
 
   cf_options.blob_garbage_collection_force_threshold = 1.5;
-  ASSERT_TRUE(ColumnFamilyData::ValidateOptions(db_options, cf_options)
-                  .IsInvalidArgument());
+  ASSERT_TRUE(ValidateOptions(db_options, cf_options).IsInvalidArgument());
 }
 
 }  // namespace ROCKSDB_NAMESPACE
