+++ db/db_impl/db_impl.h
@@ -1914,12 +1914,6 @@ class DBImpl : public DB {
   IOStatus CreateWAL(uint64_t log_file_num, uint64_t recycle_log_number,
                      size_t preallocate_block_size, log::Writer** new_log);
 
-  // Validate self-consistency of DB options
-  static Status ValidateOptions(const DBOptions& db_options);
-  // Validate self-consistency of DB options and its consistency with cf options
-  static Status ValidateOptions(
-      const DBOptions& db_options,
-      const std::vector<ColumnFamilyDescriptor>& column_families);
 
   // Utility function to do some debug validation and sort the given vector
   // of MultiGet keys
